test: fix necessary `getExperimentVariant` mocks
